Strongly improve AMQP error manegement
This commit strongly improves how the relayer manages errors
occurring within the AMQP client.
The relayer will now properly terminate if an error occurs
in the AMQP connection (e.g., if a broken connection goes down),
or it will re-attempt a connection after a given number of
seconds if the new option -R/--retry-interval is specified.
This commit introduces also an additional utility module (Timers),
to easily create periodic timers.
It is currently unused, but it may be very useful in future
versions of the relayer.